Some tweaks to bt-messages

+++ src/se/bengtsson/thegame/game/controller/Controller.java
@@ -6,5 +6,5 @@ public interface Controller {
 
 	public boolean isRightTriggerPressed();
 
-	public float getTilt();
+	public byte getTilt();
 }
